refactor: context organization to improve global and local context management
